+++ lib/core/background_task/task_manager.dart
@@ -1,8 +1,13 @@
-import 'package:aniflow/core/background_task/task.dart';
+import 'package:aniflow/core/background_task/tasks/send_notification_task.dart';
+import 'package:aniflow/core/background_task/tasks/task.dart';
 import 'package:aniflow/core/data/auth_repository.dart';
 import 'package:collection/collection.dart';
 import 'package:injectable/injectable.dart';
 import 'package:permission_handler/permission_handler.dart';
+
+mixin BackgroundTaskName {
+  static const sendNotificationTaskName = 'post_platform_notifications_task_2';
+}
 
 @injectable
 class BackgroundTaskFactory {
@@ -10,15 +15,15 @@ class BackgroundTaskFactory {
 
   final AuthRepository _authRepository;
 
-  Future<List<Task>> createTasks() async {
-    List<Task?> tasks = await Future.wait([
+  Future<List<BackgroundTask>> createTasks() async {
+    List<BackgroundTask?> tasks = await Future.wait([
       _getSendNotificationTaskOrNull(),
     ]);
 
     return tasks.whereNotNull().toList();
   }
 
-  Future<Task?> _getSendNotificationTaskOrNull() async {
+  Future<BackgroundTask?> _getSendNotificationTaskOrNull() async {
     final authedUser = await _authRepository.getAuthedUser();
     final permissionStatus = await Permission.notification.status;
     if (authedUser != null && permissionStatus == PermissionStatus.granted) {
